Move decoder and test to module
